move CachedBestsellingProductFacade::getAllOfferedBestsellingProductIds() from project-base to the framework package
- removed deprecated getAllOfferedBestsellingProducts()
